+++ framework/shared/components/blade-navigation/components/vc-blade-navigation/vc-blade-navigation.vue
@@ -1,46 +1,76 @@
 <template>
-  <router-view v-slot="{ Component, route }">
-    <component
-      :is="Component"
-      :closable="false"
-      v-show="$isMobile.value ? !blades.length : blades.length <= 1"
-      @open:blade="onBladeOpen($event, 0)"
-      :options="parentBladeOptions"
-      :expanded="blades.length === 0"
-      :maximized="findStateById(0)"
-      :blades="blades"
-      :param="resolveParam"
-      :key="route"
-      :ref="(el: IBladeElement) => setParentRef(el, Component)"
-      @expand:blade="handleMaximizeBlade(0, true)"
-      @collapse:blade="handleMaximizeBlade(0, false)"
+  <ErrorInterceptor
+    capture
+    v-slot="{ error, reset }"
+  >
+    <router-view
+      @vnode-before-unmount="reset"
+      v-slot="{ Component, route }"
+      :key="route.path"
     >
-    </component>
-  </router-view>
-  <component
+      <component
+        :is="Component"
+        :closable="false"
+        v-show="$isMobile.value ? !blades.length : blades.length <= 1"
+        @open:blade="onBladeOpen($event, 0)"
+        :options="parentBladeOptions"
+        :expanded="blades.length === 0"
+        :maximized="findStateById(0)"
+        :blades="blades"
+        :param="resolveParam"
+        :key="route.path"
+        :ref="(el: IBladeElement) => setParentRef(el, Component)"
+        @expand:blade="handleMaximizeBlade(0, true)"
+        @collapse:blade="handleMaximizeBlade(0, false)"
+      >
+        <template
+          v-slot:error
+          v-if="error"
+          >{{ error }}</template
+        >
+      </component>
+    </router-view>
+  </ErrorInterceptor>
+
+  <template
     v-for="(blade, i) in blades"
-    v-show="i >= blades.length - ($isMobile.value ? 1 : 2)"
     :key="`blade_${i + 1}`"
-    :is="blade.component"
-    :param="blade.param"
-    :closable="i >= 0"
-    :expanded="i === blades.length - 1"
-    :maximized="findStateById(blade.idx)"
-    :options="blade.bladeOptions"
-    @open:blade="onBladeOpen($event, blade.idx)"
-    @close:blade="onBladeClose(i)"
-    @close:children="$emit('onClose', i + 1)"
-    @parent:call="$emit('onParentCall', { id: i, args: $event })"
-    @expand:blade="handleMaximizeBlade(blade.idx, true)"
-    @collapse:blade="handleMaximizeBlade(blade.idx, false)"
-    :ref="(el: IBladeElement) => setBladesRef(el, blade)"
-  ></component>
+  >
+    <ErrorInterceptor
+      v-slot="{ error }"
+      capture
+    >
+      <component
+        v-show="i >= blades.length - ($isMobile.value ? 1 : 2)"
+        :is="blade.component"
+        :param="blade.param"
+        :closable="i >= 0"
+        :expanded="i === blades.length - 1"
+        :maximized="findStateById(blade.idx)"
+        :options="blade.bladeOptions"
+        @open:blade="onBladeOpen($event, blade.idx)"
+        @close:blade="onBladeClose(i)"
+        @close:children="$emit('onClose', i + 1)"
+        @parent:call="$emit('onParentCall', { id: i, args: $event })"
+        @expand:blade="handleMaximizeBlade(blade.idx, true)"
+        @collapse:blade="handleMaximizeBlade(blade.idx, false)"
+        :ref="(el: IBladeElement) => setBladesRef(el, blade)"
+      >
+        <template
+          v-slot:error
+          v-if="error"
+          >{{ error }}</template
+        >
+      </component>
+    </ErrorInterceptor>
+  </template>
 </template>
 
 <script lang="ts" setup>
 import { computed, ref, VNode } from "vue";
 import { useRoute } from "vue-router";
 import { IBladeContainer, IBladeElement, IBladeEvent, IParentCallArgs, IBladeRef } from "./../../../../../shared";
+import { ErrorInterceptor } from "./../../../../../core/plugins/error-interceptor";
 
 export interface Props {
   blades: IBladeContainer[];