+++ framework/shared/components/blade-navigation/composables/useBladeNavigation/index.ts
@@ -110,8 +110,6 @@ export function useBladeNavigation(): IUseBladeNavigation {
           blades.value[index]?.onClose?.();
         }
         blades.value.splice(index);
-      } else {
-        throw "Closing prevented";
       }
     }
   }