+++ framework/shared/components/notifications/components/index.ts
@@ -1,0 +1,1 @@
+export { default as NotificationContainer } from "./notification-container";