+++ framework/ui/components/molecules/vc-notification/index.ts
@@ -1,10 +1,3 @@
-import { VNode } from "vue";
 import _Notification from "./vc-notification.vue";
 
-export const VcNotification = _Notification as typeof _Notification & {
-  new (): {
-    $slots: {
-      default: () => VNode[];
-    };
-  };
-};
+export const VcNotification = _Notification as typeof _Notification;